+++ src/pages/learn/workshops.js
@@ -1,16 +1,16 @@
 import React from "react";
 import { ThemeProvider } from "styled-components";
 
-import Layout from "../components/layout";
-import SEO from "../components/seo";
+import Layout from "../../components/layout";
+import SEO from "../../components/seo";
 
-import Navigation from "../sections/Navigation";
-import WorkshopsPage from "../sections/Workshop-grid";
-import Footer from "../sections/Footer";
+import Navigation from "../../sections/Navigation";
+import WorkshopsPage from "../../sections/Workshop-grid";
+import Footer from "../../sections/Footer";
 
-import { GlobalStyle } from "../sections/app.style";
-import theme from "../theme/blog/themeStyles";
-import WorkshopBottom from "../components/Workshop-Bottom";
+import { GlobalStyle } from "../../sections/app.style";
+import theme from "../../theme/blog/themeStyles";
+import WorkshopBottom from "../../components/Workshop-Bottom";
 
 
 
@@ -19,7 +19,7 @@ const WorkshopsGridPage = ({hide_path}) => {
         <ThemeProvider theme={theme}>
             <Layout>
                 <GlobalStyle />
-                <SEO title="Workshops" />
+                <SEO title="Service Mesh Workshops" />
                 <Navigation />
                 <WorkshopsPage hide_path={hide_path} />
                 <WorkshopBottom />
